chore: Refactor BrushStyles types to own file.
