+++ packages/spec/src/spec/interactors/BrushStyles.ts
@@ -1,0 +1,27 @@
+/** Styles for rectangular selection brushes. */
+export interface BrushStyles {
+  /**
+   * The overall opacity of the brush rectangle.
+   */
+  opacity?: number;
+  /**
+   * The fill opacity of the brush rectangle.
+   */
+  fillOpacity?: number;
+  /**
+   * The stroke opacity of the brush rectangle.
+   */
+  strokeOpacity?: number;
+  /**
+   * The fill color of the brush rectangle.
+   */
+  fill?: string;
+  /**
+   * The stroke color of the brush rectangle.
+   */
+  stroke?: string;
+  /**
+   * The stroke dash array of the brush rectangle.
+   */
+  strokeDasharray?: string;
+}
