+++ packages/spec/src/spec/interactors/Interval2D.ts
@@ -1,5 +1,5 @@
 import { ParamRef } from '../Param.js';
-import { BrushStyles } from './Interval1D.js';
+import { BrushStyles } from './BrushStyles.js';
 
 /** Options for 2D interval interactors. */
 export interface Interval2DOptions {
@@ -28,9 +28,9 @@ export interface Interval2DOptions {
    */
   pixelSize?: number;
   /**
-   * A flag indicating if peer (sibling) marks are when cross-filtering
-   * (default `true`). If set, peer marks will not be filtered by this
-   * interactor's selection in cross-filtering setups.
+   * A flag indicating if peer (sibling) marks are excluded when
+   * cross-filtering (default `true`). If set, peer marks will not be
+   * filtered by this interactor's selection in cross-filtering setups.
    */
   peers?: boolean;
   /**
